portfolioBackend/requirements.txt
django
djangorestframework
django-cors-headers

Pillow